fix: repin bootstrap digest and drop a cast tree-sitter 0.27 made redundant
The bootstrap Cargo manifest digest in bin/ensure-binary.sh only tracks
Cargo.toml's content; this dependency bump changes it, so the digest
drifts until re-pinned. Re-run: python3 scripts/repin_bootstrap_digests.py

tree-sitter 0.27 changed Node::child_count() to return u32 directly, so
the existing `as u32` cast at tests/full_ast_completeness.rs:27 became
a clippy::unnecessary_cast hard error under -D warnings. Dropped.

+++ tests/full_ast_completeness.rs
@@ -24,7 +24,7 @@ use ai_architect_mcp::parser::Language;
 /// verification, not shared implementation.
 fn count_all_nodes_independently(node: tree_sitter::Node) -> usize {
     let mut count = 1;
-    for i in 0..node.child_count() as u32 {
+    for i in 0..node.child_count() {
         if let Some(child) = node.child(i) {
             count += count_all_nodes_independently(child);
         }
